+++ src/core/prelude/structures/sync-promise/interface.ts
@@ -15,15 +15,15 @@ export enum State {
 export type Value<T = unknown> = PromiseLike<T> | T;
 
 export interface ConstrResolveHandler<T = unknown> {
-	(value?: Value<T>): any;
+	(value?: Value<T>): AnyToIgnore;
 }
 
 export interface ConstrRejectHandler {
-	(reason?: unknown): any;
+	(reason?: unknown): AnyToIgnore;
 }
 
 export interface Executor<T = unknown> {
-	(resolve: ConstrResolveHandler<T>, reject: ConstrRejectHandler): any;
+	(resolve: ConstrResolveHandler<T>, reject: ConstrRejectHandler): AnyToIgnore;
 }
 
 export type ResolveHandler<V = unknown, R = V> = Function | ((value: V) => Value<R>);