+++ src/core/net/CHANGELOG.md
@@ -9,6 +9,13 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.?? (2021-??-??)
+
+#### :boom: Breaking Change
+
+* Renamed `browser.request` -> `browser-request`
+* Renamed `node.request` -> `node-request`
+
 ## v3.20.0 (2020-07-05)
 
 #### :rocket: New Feature